# Better rendering of multiline strings in Pretty formats