… src/Processors/Formats/Impl/PrettyCompactBlockOutputFormat.cpp
… src/Processors/Formats/Impl/PrettySpaceBlockOutputFormat.cpp
… tests/queries/0_stateless/02026_describe_include_subcolumns.reference
… tests/queries/0_stateless/03132_pretty_format_break_line.reference
… tests/queries/0_stateless/03132_pretty_format_break_line.sql