### Changelog category (leave one):

- Improvement

### Changelog entry (a user-readable short description of the changes that goes to CHANGELOG.md):
Multiline strings with border preservation and column width change

Example console output:

```
┌─x──────┬───y─┐
│ hello… │ 123 │
│ …world │     │
└────────┴─────┘
```

UPD: `…` in the position where the value cannot appear

```
   ┌─x─────┬───y─┐
1. │ hello…│ 123 │
   │…world │     │
   └───────┴─────┘
```
It also works with line numbering