@yariks5s 2024-02-13
also, try to avoid cases like
```C++
if (condition) {
```
use instead:
```C++
if (condition)
{
```