@Volodyachan 2024-02-13
> Hello! Can you please also add tests to check how your changes work?

Okay, I'll do it later, but right now there's another problem.