@Volodyachan 2024-02-13
There are 2 tests where my solution gives a different answer

Example of current answer:

```
┌─name─┬─type─────────────────────────────┬─default_type─┬─default_expression─┬─comment──────┬─codec_expression─┬─ttl_expression─┐
│ id   │ UInt64                           │              │                    │ index column │                  │                │
│ arr  │ Array(UInt64)                    │ DEFAULT      │ [10, 20]           │              │ ZSTD(1)          │                │
│ t    │ Tuple(
    a String,
    b UInt64) │ DEFAULT      │ ('foo', 0)         │              │ ZSTD(1)          │                │
└──────┴──────────────────────────────────┴──────────────┴────────────────────┴──────────────┴──────────────────┴────────────────┘
```

Example of my answer:
```
┌─name─┬─type────────────┬─default_type─┬─default_expression─┬─comment──────┬─codec_expression─┬─ttl_expression─┐
│ id   │ UInt64          │              │                    │ index column │                  │                │
│ arr  │ Array(UInt64)   │ DEFAULT      │ [10, 20]           │              │ ZSTD(1)          │                │
│ t    │ Tuple(…         │ DEFAULT      │ ('foo', 0)         │              │ ZSTD(1)          │                │
│      │ …    a String,… │              │                    │              │                  │                │
│      │ …    b UInt64)  │              │                    │              │                  │                │
└──────┴─────────────────┴──────────────┴────────────────────┴──────────────┴──────────────────┴────────────────┘
```

Can I canonize the answer, and if so, how?