@alexey-milovidov 2024-02-14
@Volodyachan, the new behavior looks much better! Yes, you should update the `.reference` file for the test.
To do it, you can run the test alone with

`clickhouse-client -n < test.sql > test.reference`