@alexey-milovidov 2024-02-14
I advise to place `…` in the position where the value cannot appear - for disambiguation:
```
┌─x─────┬───y─┐
│ hello…│ 123 │
│…world │     │
└───────┴─────┘
```