@Volodyachan 2024-04-29
@yariks5s  Do all tests have to be passed for merge? Because the dropped tests seem to have nothing to do with my code. If they shouldn't, you can merge? I don't have such opportunity)