@azat 2024-05-16
It looks like this PR makes PrettySpace format significantly slower - https://play.clickhouse.com/play?user=play#U0VMRUNUIGNoZWNrX3N0YXJ0X3RpbWUsIHRlc3RfZHVyYXRpb25fbXMsIHRlc3RfbmFtZSwgcmVwb3J0X3VybApGUk9NIGNoZWNrcwpXSEVSRSB0ZXN0X25hbWUgTElLRSAnJXNlbGVjdF9mb3JtYXQlJwogICAgQU5EIGNoZWNrX3N0YXJ0X3RpbWUgPj0gbm93KCkgLSBJTlRFUlZBTCAyMCBkYXkKICAgIGFuZCByZXBvcnRfdXJsIGxpa2UgJyUuNCcKT1JERVIgQlkgY2hlY2tfbmFtZSwgdGVzdF9uYW1lLCBjaGVja19zdGFydF90aW1l