@Algunenano 2024-05-16
Confirmed.

Before the revert:
```
stage   start   0.005   0.005
stage   parse   0.000   0.006
stage   before-connect  0.000   0.006
server  0       localhost       49000
stage   connect 0.000   0.006
drop    0       0.001971006393432617    DROP TABLE IF EXISTS table_PrettySpace
stage   drop-1  0.002   0.008
stage   settings        0.001   0.009
create  0       0.011439085006713867    CREATE TABLE IF NOT EXISTS table_PrettySpace ENGINE = File(PrettySpace, '/dev/null') AS test.hits
stage   create  0.012   0.021
stage   sync    0.021   0.042
display-name    0       INSERT INTO table_PrettySpace SELECT * FROM test.hits LIMIT 10000
prewarm 0       select_format.query0.prewarm0   0       2.6575686931610107
query   0       select_format.query0.run0       0       1.8274660110473633
client-time     0       1.8275002849986777      1.8274660110473633
profile-total   0
stage   run     4.485   4.527
drop    0       0.0006453990936279297   DROP TABLE IF EXISTS table_PrettySpace
stage   drop-2  0.001   4.527
```

After the revert:

```
/mnt/ch/ClickHouse/tests/performance $ ../../tests/performance/scripts/perf.py --runs 1 select_format.xml  --port 49000
stage   start   0.006   0.006
stage   parse   0.000   0.006
stage   before-connect  0.000   0.006
server  0       localhost       49000
stage   connect 0.000   0.006
drop    0       0.002940654754638672    DROP TABLE IF EXISTS table_PrettySpace
stage   drop-1  0.003   0.009
stage   settings        0.001   0.010
create  0       0.01272130012512207     CREATE TABLE IF NOT EXISTS table_PrettySpace ENGINE = File(PrettySpace, '/dev/null') AS test.hits
stage   create  0.013   0.023
stage   sync    0.026   0.049
display-name    0       INSERT INTO table_PrettySpace SELECT * FROM test.hits LIMIT 10000
prewarm 0       select_format.query0.prewarm0   0       0.29047489166259766
query   0       select_format.query0.run0       0       0.23655939102172852
query   0       select_format.query0.run1       0       0.23544073104858398
query   0       select_format.query0.run2       0       0.22042036056518555
query   0       select_format.query0.run3       0       0.22723174095153809
query   0       select_format.query0.run4       0       0.23174118995666504
client-time     0       1.1515622549995896      1.1513934135437012
profile-total   0
stage   run     1.442   1.492
drop    0       0.0005781650543212891   DROP TABLE IF EXISTS table_PrettySpace
stage   drop-2  0.001   1.492
```

Taking 2 seconds to show 10k lines is not good.